other.Points

+++ Lib/Builder.cs
@@ -4,12 +4,12 @@ namespace Shape.Lib;
 
 public class Builder
 {
-    public static Point Build(int x, int y)
+    public static Point Build(double x, double y)
     {
         return new Point(x, y);
     }
 
-    public static Point[] Build(params (int, int)[] points)
+    public static Point[] Build(params (double, double)[] points)
     {
         return points
             .Select(coord => Build(coord.Item1, coord.Item2))
